requirements.txt
black==23.3.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.6
distlib==0.3.6
filelock==3.12.0
identify==2.5.24
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
pytest==7.3.1
pytest-cov==4.1.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
ruff==0.0.270
urllib3==2.0.2
virtualenv==20.23.0
